requirements.txt
click==7.1.1
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
h5py==2.10.0
imageio==2.8.0
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
Werkzeug==1.0.0
